updated post editor with a fresh new look and full support for setting the tags and category
fixed tags not being auto created for new threads
